poba_api_py/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
substrate-interface==1.7.11
python-dotenv==1.0.1
apscheduler
websockets